Fix up lifecycle issues & update logger (#46)
The issue (CLI not shutting down sometimes) was that I was misusing swift service lifecycle. This cleans up those parts & fixes the issue. Don't want to get too cocky but I think it might also fix the issue with the Xcode debugger occasionally not stopping properly once and for all 🙏