Messaging (#88)
Adds a core abstraction for message channels as well as 4 concrete drivers - Slack, APNs, SendGrid, and Twilio. Heavily inspired by Laravel Notifications.

# Channel

`Channel` is a protocol that contains a message and a receiver type. These are the foundations of setting up abstractions over various methods of notifications.

`Codable` messages can optionally be queued or stored.

`Notification` represents a collection of messages to send at once, and can be dispatched to a specific receiver.